detect c header files

+++ ci/test_cmake.sh
@@ -5,13 +5,40 @@
 
 set -exu
 
+failures=()
+
 for dir in example_packages/*/ ; do
+
+	# TODO
+	[[ "$dir" == "example_packages/c_includes/" ]] && continue
+	[[ "$dir" == "example_packages/c_main/" ]] && continue
+	[[ "$dir" == "example_packages/c_main_preprocess/" ]] && continue
+	[[ "$dir" == "example_packages/circular_example/" ]] && continue
+	[[ "$dir" == "example_packages/circular_test/" ]] && continue
+	[[ "$dir" == "example_packages/dependency_priority/" ]] && continue
+	[[ "$dir" == "example_packages/features_with_dependency/" ]] && continue
+	[[ "$dir" == "example_packages/fixed-form/" ]] && continue
+	[[ "$dir" == "example_packages/metapackage_blas/" ]] && continue
+	[[ "$dir" == "example_packages/metapackage_blas/" ]] && continue
+	[[ "$dir" == "example_packages/metapackage_hdf5/" ]] && continue
+	[[ "$dir" == "example_packages/metapackage_mpi/" ]] && continue
+	[[ "$dir" == "example_packages/metapackage_mpi_c/" ]] && continue
+	[[ "$dir" == "example_packages/metapackage_mpi_cpp/" ]] && continue
+	[[ "$dir" == "example_packages/metapackage_netcdf/" ]] && continue
+	[[ "$dir" == "example_packages/metapackage_stdlib_extblas/" ]] && continue
+	[[ "$dir" == "example_packages/preprocess_cpp_c/" ]] && continue
+	[[ "$dir" == "example_packages/shared_app_only/" ]] && continue
+	[[ "$dir" == "example_packages/shared_lib_empty/" ]] && continue
+	[[ "$dir" == "example_packages/static_app_only/" ]] && continue
+	[[ "$dir" == "example_packages/static_lib_empty/" ]] && continue
 
 	pushd "$dir"
 
 	../../build/bin/fpm generate --cmake
 	cmake -B temp_cmake_build -S .
-	cmake --build temp_cmake_build --parallel
+	if [[ ! $(cmake --build temp_cmake_build --parallel) ]] ; then
+		failures+=("$dir")
+	fi
 
 	# Cleanup
 	rm CMakeLists.txt
@@ -21,5 +48,13 @@ for dir in example_packages/*/ ; do
 
 done
 
-# TODO: test fpm self-generation too
+echo "failures = "
+echo "${failures[@]}"
+
+# TODO: test fpm self-generation too:
+#
+#     fpm run -- generate --cmake
+#     cmake -B build -S .
+#     cmake --build build --parallel
+#
